requirements.txt
# Answer Evaluation Engine - Dependencies
# Used by setup.sh and the run command

# Core Web Framework
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.12

# Data Validation
pydantic==2.9.2

# OCR & Document Processing
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1
python-docx==1.1.2
pillow==10.4.0
opencv-python-headless==4.10.0.84

# AI & ML
sentence-transformers==3.1.1
scikit-learn==1.5.1
numpy==1.26.4

# Utilities
requests==2.32.3
jinja2==3.1.4
python-dotenv==1.0.1